decouple core logic from GUI
